tradegpt-backend/requirements.txt
fastapi
uvicorn
python-dotenv
pandas
numpy
pydantic
coinbase-advanced-py
requests
ta
azure-ai-inference
azure-core
azure-identity
azure-keyvault-secrets
statistics
